Renaming modules according to coding standard
