+++ backend/controllers/product-controller.js
@@ -1,6 +1,7 @@
 import { ProductModel } from '../models/product-model.js';
 import { OrderModel } from '../models/order-model.js';
 import { pool } from '../db/pool.js';
+import fs from 'fs';
 
 class ProductController {
 
@@ -184,6 +185,89 @@ class ProductController {
 
         return false;
     }
+
+
+     /**
+     * Add product, after validating inputs.
+     * @param {Object} req - The request object.
+     * @param {Object} res - The response object.
+     */
+     static async add(req, res){
+        const { productName, category, description, price } = req.body;
+
+        if(!(productName && category && price)){
+            return res.json({ message: 'All required input fields must be filled!'});
+        }
+
+        const thumbnailImgPath = req.store.galleryImgsPath.replace('gallery','product') + req.uniqueProductName + '.jpg';
+
+        try{
+            await ProductModel.add(req.params.storeId, category, productName, description, price, thumbnailImgPath);
+            return res.redirect('/farmer/store/' + req.params.storeId);
+        }catch(err) {
+            if(err.code == 'ER_DUP_ENTRY'){
+                const sqlMessageParse = /^Duplicate entry '(.*)' for key '(.*)'$/.exec(err.sqlMessage)
+
+                if(sqlMessageParse[2] == 'product_AK'){
+                    return res.json({ message: 'This store already has another product with this name!' });
+                }
+
+                return res.json({ message: 'The ' + sqlMessageParse[2] + ' enterred is in use by another product!' });
+            }
+            console.log(err);
+            res.status(500).json({ message: "Server Error" });
+        }
+    }
+
+
+    /**
+     * Update product by Id, after validating input.
+     * @param {Object} req - The request object.
+     * @param {Object} res - The response object.
+     */
+    static async update(req, res){
+        const { productName, category, description, price } = req.body;
+
+        if(!(productName && category && price)){
+            return res.json({ message: 'All required input fields must be filled!'});
+        }
+
+        try{
+            await ProductModel.update(req.params.productId, category, productName, description, price);
+            return res.redirect('/farmer/store/' + req.params.storeId);
+        }catch(err) {
+            if(err.code == 'ER_DUP_ENTRY'){
+                const sqlMessageParse = /^Duplicate entry '(.*)' for key '(.*)'$/.exec(err.sqlMessage)
+
+                if(sqlMessageParse[2] == 'product_AK'){
+                    return res.json({ message: 'This store already has another product with this name!' });
+                }
+
+                return res.json({ message: 'The ' + sqlMessageParse[2] + ' enterred is in use by another product!' });
+            }
+            console.log(err);
+            res.status(500).json({ message: "Server Error" });
+        }
+    }
+
+    /**
+     * Delete product by ID and corresponding image directories stored in server.
+     * @param {Object} req - The request object.
+     * @param {Object} res - The response object.
+     */
+    static async delete(req, res){
+        try{
+            const productImagePath = res.locals.product.thumbnailImgPath;
+            fs.rmSync(productImagePath);
+            await ProductModel.delete(req.params.productId);
+            res.redirect('/farmer/store/' + req.params.storeId);
+        }catch(err){
+            console.log(err);
+            res.status(500).json({ message: "Server Error" });
+        }
+    }
+
+
 }
 
 export {
