+++ backend/middlewares/passport/local-strategies/farmer.js
@@ -1,5 +1,5 @@
 import bcryptjs from 'bcryptjs';
-import { FarmerModel } from '../../../models/farmerModel.js';
+import { FarmerModel } from '../../../models/farmer-model.js';
 import { Strategy as LocalStrategy } from 'passport-local';
 /**
  * Authenticates a farmer using their email and password.
